Added Delete Confirm Dialog

+++ modules/ui/tabs/fileUI.lua
@@ -2,7 +2,9 @@ fileUI = {
     saveBoxSize = {x = 525, y = 80},
 	colors = {frame = {0, 50, 255}},
     filesData = {},
-    names = {}
+    names = {},
+    popup = false,
+    deleteFile = nil
 }
 
 function fileUI.drawFile(recorder, file)
@@ -31,14 +33,14 @@ function fileUI.drawFile(recorder, file)
         pressed = ImGui.Button("Apply Name")
         if pressed then 
             fileUI.filesData[name].info.name = fileUI.names[name]  -- Change name logic
-            fileUI.filesData[name].info = info 
+            fileUI.filesData[name].info = info
             recorder.config.saveFile("saves/" .. name .. ".json", fileUI.filesData[name])
             os.rename("saves/" .. name .. ".json", "saves/" .. fileUI.filesData[name].info.name .. ".json")
         end
 
         ImGui.Separator()
         ImGui.Text("Recorded on: " .. info.recordedOn .. " | Length: " .. info.frames .. " | Location: " .. info.location) -- Some small infos
-        
+
         pressed = ImGui.Button("Load")
         if pressed then 
             recorder.baseUI.switchToArrange = true
@@ -47,20 +49,54 @@ function fileUI.drawFile(recorder, file)
 
         ImGui.SameLine()
         pressed = ImGui.Button("Delete File")
-        if pressed then 
-            os.remove("saves/" .. name .. ".json")
-            fileUI.filesData[name] = nil
-        end 
-        
+        if pressed then
+            if recorder.settings.deleteConfirm then
+                fileUI.popup = true
+                fileUI.deleteFile = file
+            else
+                os.remove("saves/" .. name .. ".json")
+                fileUI.filesData[name] = nil
+            end
+        end
+
         ImGui.EndChild()
         recorder.CPS.colorEnd(2)
     end
 end
 
+function fileUI.handlePopUp(recorder) -- ToDo : Make this its own universal module
+    if fileUI.popup then
+        ImGui.OpenPopup("Delete Record?")
+        if ImGui.BeginPopupModal("Delete Record?", true, ImGuiWindowFlags.AlwaysAutoResize) then
+            local again, changed = ImGui.Checkbox("Dont ask again", not recorder.settings.deleteConfirm)
+            if changed then recorder.settings.deleteConfirm = not again end
+
+            if ImGui.Button("Cancel") then
+                ImGui.CloseCurrentPopup()
+                fileUI.popup = false
+            end
+
+            ImGui.SameLine()
+
+            if ImGui.Button("Confirm") then
+                ImGui.CloseCurrentPopup()
+                os.remove("saves/" .. fileUI.deleteFile.name)
+                fileUI.filesData[fileUI.deleteFile.name] = nil
+                fileUI.popup = false
+            end
+            ImGui.EndPopup()
+        end
+    end
+end
+
 function fileUI.draw(recorder)
+    if next(dir("saves")) == nil then
+        ImGui.Text("Nothing here ... Record something first!")
+    end
     for _, file in pairs(dir("saves")) do
         fileUI.drawFile(recorder, file)
-    end 
+    end
+    fileUI.handlePopUp(recorder)
 end
 
 return fileUI
